requirements.txt
Django==2.0.4
Pillow==5.1.0
psycopg2==2.7.4
python-memcached==1.59
pytz==2018.4
six==1.11.0
sorl-thumbnail==12.4.1
uWSGI==2.0.17
